requirements.txt
streamlit
PyPDF2
faiss-cpu
torch
sentence-transformers
google-generativeai
numpy
python-dotenv
